backend/requirements.txt
flask==2.3.3
flask-socketio==5.3.6
flask-cors==4.0.0
python-dotenv==1.0.0
numpy>=1.24.0
openai>=1.0.0
webrtcvad>=2.0.10
gunicorn==20.1.0

# Optional async backends, install based on your preference
# For Python < 3.13, you can uncomment one of these:
# eventlet==0.33.3
# gevent==24.2.1
# gevent-websocket==0.10.1 